requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
aiohttp==3.6.2
aioredis==1.3.1
aiosqlite==0.11.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
ciso8601==2.1.2
hiredis==1.0.1
idna==2.8
multidict==4.7.3
pypika==0.35.18
redis==3.3.11
toml==0.10.0
tortoise-orm==0.15.4
typing-extensions==3.7.4.1
yarl==1.4.2
